, d36ee881012019b1d884cef3a80df56064cfb8ea

 and c5f1f461bf594cabe540d4a99cff793ceb3a59eb

 - wcms bugfixes)

git-svn-id: https://server.mycore.de/svn/mycore/branches/mycore-2014.06.0.x@30759 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-wcms2/src/main/java/org/mycore/wcms2/navigation/MCRWCMSContentManager.java
@@ -18,6 +18,7 @@ import org.jdom2.Element;
 import org.jdom2.JDOMException;
 import org.jdom2.input.SAXBuilder;
 import org.jdom2.output.Format;
+import org.jdom2.output.Format.TextMode;
 import org.jdom2.output.XMLOutputter;
 import org.mycore.common.config.MCRConfiguration;
 import org.mycore.common.content.MCRURLContent;
@@ -141,7 +142,7 @@ public class MCRWCMSContentManager {
      * @param items
      */
     public void save(JsonArray items) {
-        XMLOutputter out = new XMLOutputter(Format.getPrettyFormat().setEncoding("UTF-8"));
+        XMLOutputter out = new XMLOutputter(Format.getCompactFormat().setEncoding("UTF-8"));
         for (JsonElement e : items) {
             if (!e.isJsonObject()) {
                 LOGGER.warn("Invalid json element in items " + e);
@@ -180,16 +181,27 @@ public class MCRWCMSContentManager {
 
     public void move(String from, String to) {
         try {
-            // copy
+            // get from
             URL fromURL = MCRWebPagesSynchronizer.getURL(from);
-            SAXBuilder builder = new SAXBuilder();
-            Document document = builder.build(fromURL);
+            Document document;
+            if (fromURL == null) {
+                // if the from resource couldn't be found we assume its not created yet.
+                MyCoReWebPageProvider wpp = new MyCoReWebPageProvider();
+                wpp.addSection("neuer Eintrag", new Element("p").setText("TODO"), "de");
+                document = wpp.getXML();
+            } else {
+                SAXBuilder builder = new SAXBuilder();
+                document = builder.build(fromURL);
+            }
+            // save
             XMLOutputter out = new XMLOutputter(Format.getPrettyFormat().setEncoding("UTF-8"));
             try (OutputStream fout = MCRWebPagesSynchronizer.getOutputStream(to)) {
                 out.output(document, fout);
             }
             // delete old
-            Files.delete(Paths.get(fromURL.toURI()));
+            if (fromURL != null) {
+                Files.delete(Paths.get(fromURL.toURI()));
+            }
         } catch (Exception exc) {
             LOGGER.error("Error moving " + from + " to " + to, exc);
             throwError(ErrorType.couldNotMove, to);
